# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AverNoxTrader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthInput {
  username: string;
  password: string;
  /** @nullable */
  referralCode?: string | null;
}

export interface User {
  id: number;
  username: string;
  createdAt: string;
  referralCode?: string;
  accountStatus?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Balance {
  total: number;
  tradingBalance: number;
  profitReserve: number;
  emergencyReserve: number;
  withdrawableBalance: number;
  historicalProfit: number;
}

export interface UserSettings {
  id: number;
  userId: number;
  tradingStyle?: string;
  executionMode?: string;
  maxActivePositions?: number;
  smartEntry?: boolean;
  riskLevel?: number;
  stopLossProtection?: boolean;
  profitLock?: boolean;
  capitalExposureLimit?: number;
  confluenceStrength?: number;
  marketScanner?: boolean;
  aiAdaptation?: string;
  signalQuality?: string;
  allocationMethod?: string;
  reserveCapital?: number;
  autoRebalancing?: boolean;
  tradeNotifications?: boolean;
  performanceReports?: string;
  dashboardMode?: string;
}

export interface UserSettingsUpdate {
  tradingStyle?: string;
  executionMode?: string;
  maxActivePositions?: number;
  smartEntry?: boolean;
  riskLevel?: number;
  stopLossProtection?: boolean;
  profitLock?: boolean;
  capitalExposureLimit?: number;
  confluenceStrength?: number;
  marketScanner?: boolean;
  aiAdaptation?: string;
  signalQuality?: string;
  allocationMethod?: string;
  reserveCapital?: number;
  autoRebalancing?: boolean;
  tradeNotifications?: boolean;
  performanceReports?: string;
  dashboardMode?: string;
}

export interface ChangePasswordInput {
  oldPassword: string;
  newPassword: string;
}

export interface DepositInput {
  amount: number;
  transactionCode: string;
  method?: string;
}

export interface Deposit {
  id: number;
  userId: number;
  amount: number;
  transactionCode: string;
  method?: string;
  status: string;
  createdAt: string;
}

export interface DailySnapshot {
  day: number;
  value: number;
  profit: number;
}

export interface Investment {
  id: number;
  userId: number;
  principal: number;
  currentValue: number;
  startDate: string;
  endDate: string;
  status: string;
  dayNumber: number;
  profitPercent?: number;
  dailySnapshots?: DailySnapshot[];
}

export interface GrowthPoint {
  date: string;
  value: number;
}

export interface InvestmentSummary {
  totalProfit: number;
  monthlyReturn: number;
  avgTrade: number;
  winRate: number;
  lossRate: number;
  totalWithdrawn: number;
  bestStrategy: string;
  growthCurve: GrowthPoint[];
}

export interface Trade {
  id: number;
  investmentId: number;
  asset: string;
  type: string;
  amount: number;
  result: string;
  profitLoss: number;
  executedAt: string;
  status: string;
  strategy?: string;
}

export interface WithdrawInput {
  amount: number;
  crypto: string;
  walletAddress: string;
}

export interface Withdrawal {
  id: number;
  userId: number;
  amount: number;
  crypto: string;
  walletAddress: string;
  status: string;
  createdAt: string;
  cryptoAmount?: number;
  /** @nullable */
  reversalReason?: string | null;
}

export interface Asset {
  symbol: string;
  name: string;
  price: number;
  change: number;
  changePercent: number;
  volume: string;
  category?: string;
}

export interface TrendingAssets {
  gainers: Asset[];
  losers: Asset[];
}

export interface TradeAlert {
  id: number;
  title: string;
  description: string;
  impact: string;
  category: string;
  publishedAt: string;
  assets?: string[];
}

export interface CryptoRate {
  symbol: string;
  name: string;
  usdRate: number;
  logoUrl: string;
}

export interface Treasury {
  tradingBalance: number;
  profitReserve: number;
  emergencyReserve: number;
  withdrawableBalance: number;
  historicalProfit: number;
  reinvestPercent: number;
  reservePercent: number;
  withdrawablePercent: number;
  compoundingRate: number;
  compoundingMode: string;
}

export interface TreasurySettingsInput {
  reinvestPercent?: number;
  reservePercent?: number;
  withdrawablePercent?: number;
  compoundingRate?: number;
  compoundingMode?: string;
}

export interface Portfolio {
  cryptoPercent: number;
  forexPercent: number;
  stocksPercent: number;
  experimentalPercent: number;
  totalValue: number;
}

export interface PortfolioInput {
  cryptoPercent?: number;
  forexPercent?: number;
  stocksPercent?: number;
  experimentalPercent?: number;
}

export interface ReferralEntry {
  id: number;
  username: string;
  joinedAt: string;
  reward: number;
}

export interface Referral {
  code: string;
  link: string;
  totalReferrals: number;
  totalEarned: number;
  referralReward: number;
  history?: ReferralEntry[];
}

export interface Goal {
  id: number;
  userId: number;
  title: string;
  targetAmount: number;
  currentAmount: number;
  deadline: string;
  status: string;
  actionOnCompletion?: string;
}

export interface GoalInput {
  title: string;
  targetAmount: number;
  deadline: string;
  actionOnCompletion?: string;
}

export interface SupportIssue {
  id: number;
  title: string;
  category: string;
  /** @nullable */
  description?: string | null;
}

export interface SupportRequestInput {
  issueId: number;
  description?: string;
}

export interface SupportRequest {
  id: number;
  issueId: number;
  status: string;
  createdAt: string;
}