# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AverNoxTrader API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /user/me:
    get:
      operationId: getMe
      tags: [user]
      summary: Get current user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /user/settings:
    get:
      operationId: getUserSettings
      tags: [settings]
      summary: Get user settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
    patch:
      operationId: updateUserSettings
      tags: [settings]
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSettingsUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
  /user/change-password:
    post:
      operationId: changePassword
      tags: [user]
      summary: Change user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: OK
        "400":
          description: Old password does not match
...
  /portfolio:
    get:
      operationId: getPortfolio
      tags: [user]
      summary: Get user portfolio allocation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
    patch:
      operationId: updatePortfolio
      tags: [user]
      summary: Update portfolio allocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
...
  /goals:
    get:
      operationId: getGoals
      tags: [user]
      summary: Get user trading goals
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [user]
      summary: Create a new trading goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
...
components:
  schemas:
...
    AuthInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        referralCode:
          type: ["string", "null"]
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, username, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        createdAt:
          type: string
        referralCode:
          type: string
        accountStatus:
          type: string
...
    UserSettings:
      type: object
      required: [id, userId]
      properties:
        id:
          type: integer
        userId:
          type: integer
        tradingStyle:
          type: string
        executionMode:
          type: string
        maxActivePositions:
          type: integer
        smartEntry:
          type: boolean
        riskLevel:
          type: integer
        stopLossProtection:
          type: boolean
        profitLock:
          type: boolean
        capitalExposureLimit:
          type: integer
        confluenceStrength:
          type: integer
        marketScanner:
          type: boolean
        aiAdaptation:
          type: string
        signalQuality:
          type: string
        allocationMethod:
          type: string
        reserveCapital:
          type: integer
        autoRebalancing:
          type: boolean
        tradeNotifications:
          type: boolean
        performanceReports:
          type: string
        dashboardMode:
          type: string
    UserSettingsUpdate:
      type: object
      properties:
        tradingStyle:
          type: string
        executionMode:
          type: string
        maxActivePositions:
          type: integer
        smartEntry:
          type: boolean
        riskLevel:
          type: integer
        stopLossProtection:
          type: boolean
        profitLock:
          type: boolean
        capitalExposureLimit:
          type: integer
        confluenceStrength:
          type: integer
        marketScanner:
          type: boolean
        aiAdaptation:
          type: string
        signalQuality:
          type: string
        allocationMethod:
          type: string
        reserveCapital:
          type: integer
        autoRebalancing:
          type: boolean
        tradeNotifications:
          type: boolean
        performanceReports:
          type: string
        dashboardMode:
          type: string
    ChangePasswordInput:
      type: object
      required: [oldPassword, newPassword]
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
...
    Portfolio:
      type: object
      required: [cryptoPercent, forexPercent, stocksPercent, experimentalPercent, totalValue]
      properties:
        cryptoPercent:
          type: integer
        forexPercent:
          type: integer
        stocksPercent:
          type: integer
        experimentalPercent:
          type: integer
        totalValue:
          type: number
    PortfolioInput:
      type: object
      properties:
        cryptoPercent:
          type: integer
        forexPercent:
          type: integer
        stocksPercent:
          type: integer
        experimentalPercent:
          type: integer
...
    Goal:
      type: object
      required: [id, userId, title, targetAmount, currentAmount, deadline, status]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: string
        status:
          type: string
        actionOnCompletion:
          type: string
    GoalInput:
      type: object
      required: [title, targetAmount, deadline]
      properties:
        title:
          type: string
        targetAmount:
          type: number
        deadline:
          type: string
        actionOnCompletion:
          type: string
...