# lib/db/src/schema/portfolio.ts
import { pgTable, serial, integer, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const portfoliosTable = pgTable("portfolios", {
  id: serial("id").primaryKey(),
  userId: integer("user_id").notNull().unique(),
  cryptoPercent: integer("crypto_percent").notNull().default(40),
  forexPercent: integer("forex_percent").notNull().default(30),
  stocksPercent: integer("stocks_percent").notNull().default(20),
  experimentalPercent: integer("experimental_percent").notNull().default(10),
  updatedAt: timestamp("updated_at").notNull().defaultNow(),
});

export const insertPortfolioSchema = createInsertSchema(portfoliosTable).omit({ id: true, updatedAt: true });
export type InsertPortfolio = z.infer<typeof insertPortfolioSchema>;
export type Portfolio = typeof portfoliosTable.$inferSelect;